Moving argument handler to its own module.
